# docs/swagger.yaml
basePath: /v1
definitions:
  web.APIError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
host: localhost:8899
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /testapi/get-string-by-int/{some_id}:
    get:
      consumes:
      - application/json
      - application/json
      description: |-
        get string by ID
        get string by ID
      parameters:
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
      summary: Add a new pet to the store
  /testapi/get-struct-array-by-string/{some_id}:
    get:
      consumes:
      - application/json
      description: get struct array by ID
      parameters:
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Offset
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
  /user/login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口必须username,password 或 email,password
      parameters:
      - description: username
        in: formData
        name: username
        type: string
      - description: email
        in: formData
        name: email
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: token":"","expired":3600,"id":1,"uuid":"","username":"","email":""}  //成功返回的数据结构，
            最后是示例
          schema:
            type: string
        "400":
          description: code":303,"message":""}
          schema:
            type: string
      summary: 用户登录接口
      tags:
      - 用户信息   //swagger API分类标签
      - 同一个tag为一组
  /user/register:
    post:
      consumes:
      - multipart/form-data
      description: 注册接口必须 username,email,mobile, iso, password
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: mobile
        in: formData
        name: mobile
        required: true
        type: string
      - description: iso
        in: formData
        name: iso
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: id":1,"uuid":"","username":"","email":"","password":"","expired":3600}  //成功返回的数据结构，
            最后是示例
          schema:
            type: string
        "400":
          description: code":304,"message":""}
          schema:
            type: string
      summary: 用户注册接口
      tags:
      - 用户信息   //swagger API分类标签
      - 同一个tag为一组
swagger: "2.0"